Move SampleObserver out
